Break out fs.db into types
