+++ code/fs.db.types/README.md
@@ -1,0 +1,10 @@
+![Module](https://img.shields.io/badge/%40platform-fs.db.types-%23EA4E7E.svg)
+[![License: MIT](https://img.shields.io/badge/license-MIT-blue.svg)](https://opensource.org/licenses/MIT)
+[![NPM](https://img.shields.io/npm/v/@platform/fs.db.types.svg?colorB=blue&style=flat)](https://www.npmjs.com/package/@platform/fs.db.types)
+![banner](https://uih.sfo2.digitaloceanspaces.com/%40platform/repo-banners/fs.db.types.png)
+
+Common [types](https://www.typescriptlang.org) for a file-system like JSON document database.
+
+## Setup
+
+    yarn add @platform/fs.db.types
