+++ code/fs.db/src/FileDb/FileDb.TEST.ts
@@ -168,9 +168,6 @@ describe('FileDb (file-system)', () => {
     expect(events[2].payload.value).to.eql(123);
     expect(events[3].payload.value).to.eql(undefined);
     expect(events[4].payload.value).to.eql(undefined);
-
-    const get = events[0] as t.IDbGetEvent;
-    expect(get.payload.cached).to.eql(false);
   });
 
   describe('many', () => {
@@ -346,12 +343,6 @@ describe('FileDb (file-system)', () => {
 
       expect(events[4].type).to.eql('DOC/cache');
       expect(events[4].payload.action).to.eql('REMOVED');
-
-      const get1 = events[1] as t.IDbGetEvent;
-      const get2 = events[2] as t.IDbGetEvent;
-
-      expect(get1.payload.cached).to.eql(false);
-      expect(get2.payload.cached).to.eql(true);
     });
   });
 });
